pyproject.toml
[project]
name = "resume-customizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "dotenv>=0.9.9",
    "google-genai>=1.31.0",
]
